+++ ansiblediscover/entity/role/__init__.py
@@ -1,7 +1,7 @@
 import logging
 import os
 import re
-from typing import Any, Dict, List, Optional, Iterable
+from typing import List, Optional
 
 from ansiblediscover.entity import EntityFactoryABC
 from ansiblediscover.utils.fs import FS
@@ -100,6 +100,7 @@ class Meta:
         return [RoleFactory(role) for role in roles]
 
 
+# TODO This should actually be Role
 class Tasks:
     def __init__(self, content, role_path):
         self._content = content
@@ -191,67 +192,6 @@ class Tasks:
         return list(dependencies)
 
 
-class Task:
-    TASK_INCLUDE_STATEMENTS = {'include', 'include_tasks', 'import_tasks'}
-    ROLE_INCLUDE_STATEMENTS = {'include_role', 'import_role'}
-
-    @staticmethod
-    def dependencies(task: Dict[str, Any]) -> List['EntityFactoryABC']:
-        if type(task) != dict or len(task) == 0:
-            return []
-
-        dependencies = set()
-
-        if 'block' in task:
-            for block_task in task['block']:
-                dependency = Task.dependency(block_task)
-                if dependency is not None:
-                    dependencies.add(dependency)
-        else:
-            dependency = Task.dependency(task)
-            if dependency is not None:
-                dependencies.add(dependency)
-
-        return list(dependencies)
-
-    @staticmethod
-    def dependency(task: Dict[str, Any]) -> Optional['EntityFactoryABC']:
-        # TODO Ignore includes within roles for now.
-        #      We currently only handle the simple case of task include in role means a file from
-        #      within the role's tasks/ directory.
-        #
-        # task_include_stmt = Task.first_matching_include_stmt(task, Task.TASK_INCLUDE_STATEMENTS)
-        # if task_include_stmt:
-        #     name = task[task_include_stmt]
-        #
-        #     if type(name) != str:
-        #         logger.warning('task include argument is of unexpected format/type: {}'.format(task))
-        #     elif not name.endswith('.yml'):
-        #         logger.warning('task include has to have a .yml suffix, got: {}'.format(name))
-        #     else:
-        #         from ansiblediscover.entity.tasklist import TasklistFactory
-        #         return TasklistFactory(name)
-        #
-        #     return
-
-        role_include_stmt = Task.any_matching_key(task, Task.ROLE_INCLUDE_STATEMENTS)
-        if role_include_stmt:
-            include = task[role_include_stmt]
-
-            if type(include) != dict:
-                logger.warning('role include arguments are of unexpected format/type: {}'.format(task))
-            elif 'name' not in include:
-                logger.warning('role include is missing expected argument "name": {}'.format(task))
-            else:
-                return RoleFactory(include['name'], include.get('tasks_from', 'main.yml'))
-
-    @staticmethod
-    def any_matching_key(task: Dict[str, Any], keys: Iterable[str]) -> Optional[str]:
-        for key in keys:
-            if key in task:
-                return key
-
-
 class RoleFactory(EntityFactoryABC):
     def __init__(self, name: str, main_task: str = 'main.yml'):
         self._name = name
@@ -279,3 +219,6 @@ class RoleFactory(EntityFactoryABC):
 
     def build(self):
         return Role.build(self.path)
+
+
+from ansiblediscover.entity.task import Task