+++ ansiblediscover/entity/tasklist.py
@@ -1,10 +1,28 @@
+import logging
 import os
+from typing import Any, Dict, List
 
 from ansiblediscover.entity import EntityFactoryABC
+from ansiblediscover.utils.fs import FS
+
+logger = logging.getLogger(__name__)
 
 
 class Tasklist:
-    pass
+    def __init__(self, content: List[Dict[str, Any]], path: str) -> 'Tasklist':
+        self.content = content
+        self.path = path
+
+    @staticmethod
+    def build(path: str):
+        try:
+            content = FS.load_yaml(path)
+            return Tasklist(content, path)
+        except (OSError, TypeError) as e:
+            logger.error(e)
+
+    def dependencies(self) -> List['EntityFactoryABC']:
+        return [dependency for task in self.content for dependency in Task.dependencies(task)]
 
 
 class TasklistFactory(EntityFactoryABC):
@@ -24,4 +42,6 @@ class TasklistFactory(EntityFactoryABC):
         return 'tasklist'
 
     def build(self) -> 'Tasklist':
-        pass
+        return Tasklist.build(self.path)
+
+from ansiblediscover.entity.task import Task