+++ test/unit/utils/test_fs.py
@@ -39,3 +39,20 @@ def test_load_yaml_succeeds_on_valid_yaml(tmpdir):
     task_file.write(yaml.dump(content))
 
     assert content == FS.load_yaml(str(task_file))
+
+
+def test_load_yaml_succeeds_on_tags(tmpdir):
+    def default_ctor(loader, tag_suffix, node):
+        return tag_suffix + ' ' + node.value
+
+    yaml.add_multi_constructor('', default_ctor)
+    content = yaml.load("""\
+        data_contains: !vault |
+            $ANSIBLE_VAULT;1.1;AES256
+            1234567890ABCDEFGHIJKLMNO
+    """)
+
+    task_file = tmpdir.join("mydata.yml")
+    task_file.write(yaml.dump(content))
+
+    assert content == FS.load_yaml(str(task_file))